requirements.txt
streamlit== 1.43.1
Pillow==11.1.0
torch==2.7.0+cu118
torchvision== 0.22.0+cu118
